Add setAmount temporarily func for easier testing

+++ bridge/contracts/zrx_trade.sol
@@ -7,11 +7,10 @@ import "./trade_utils.sol";
 import "./IERC20.sol";
 
 contract WETH {
-    function withdraw(uint wad) public;
+    function withdraw(uint256 wad) public;
 }
 
-contract ZRXTrade is TradeUtils
-{
+contract ZRXTrade is TradeUtils {
     using LibBytes for bytes;
 
     function() external payable {}
@@ -19,24 +18,34 @@ contract ZRXTrade is TradeUtils
     address zeroProxy;
     address wETH;
 
-    constructor (address _wETH, address _zeroProxy, address payable _incognitoSmartContract)
-        public
-    {
+    constructor(
+        address _wETH,
+        address _zeroProxy,
+        address payable _incognitoSmartContract
+    ) public {
         zeroProxy = _zeroProxy;
         incognitoSmartContract = _incognitoSmartContract;
         wETH = _wETH;
     }
 
-    function trade(IERC20 srcToken, uint amount, IERC20 destToken, bytes memory callDataHex, address _forwarder) public payable isIncognitoSmartContract returns (address, uint) {
+    function trade(
+        IERC20 srcToken,
+        uint256 amount,
+        IERC20 destToken,
+        bytes memory callDataHex,
+        address _forwarder
+    ) public payable isIncognitoSmartContract returns (address, uint256) {
         // do approve if srcToken is not ETH
         approve(srcToken, zeroProxy, amount);
 
         // trigger 0x forwarder.
         IForwarder forwarder = IForwarder(_forwarder);
-        (bool success, ) = address(forwarder).call.value(msg.value)(callDataHex);
+        (bool success, ) = address(forwarder).call.value(msg.value)(
+            callDataHex
+        );
         require(success);
 
-        uint sentAmount;
+        uint256 sentAmount;
         // if destToken is ETH_CONTRACT_ADDRESS then unwrap WETH back to ETH
         if (destToken == ETH_CONTRACT_ADDRESS) {
             sentAmount = getBalance(ETH_CONTRACT_ADDRESS);
@@ -49,14 +58,14 @@ contract ZRXTrade is TradeUtils
         return (address(destToken), sentAmount);
     }
 
-    function getBalance(IERC20 token) internal view returns (uint) {
+    function getBalance(IERC20 token) internal view returns (uint256) {
         if (token == ETH_CONTRACT_ADDRESS) {
             return balanceOf(IERC20(wETH));
         }
         return balanceOf(token);
     }
 
-    function withdrawWrapETH(uint amount) public {
+    function withdrawWrapETH(uint256 amount) public {
         WETH(wETH).withdraw(amount);
     }
 }
